move more setup functions
